get user from token

+++ src/main/java/com/zwash/pojos/User.java
@@ -32,12 +32,12 @@ public class User {
 		this.dateOfBirth=dateOfBirth;
 	}
 
-	public int getId()
+	public Long getId()
 	{
 		return id;
 	}
 	
-	public void setId(int id)
+	public void setId(Long id)
 	{
 		this.id=id;
 	}
@@ -106,7 +106,7 @@ public class User {
 	@Id
 	 @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "user_seq_gen")
      @Column(name = "id", unique = true, nullable = false)
-	private int id;
+	private Long id;
 	 @Column(name = "firstName")
 	private String firstName;
 	 @Column(name = "lastName")
